requirements.txt
amqp==5.0.2
billiard==3.6.3.0
celery==5.0.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
docker==4.4.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gitdb==4.0.5
GitPython==3.1.11
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
pytz==2020.4
requests==2.25.1
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.20
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.4.0
